Added tests to verify 'Page Should Contain' keyword works with both frames and iframes. Currently iframe is failing as expected since patch has not been applied.
Added test demonstrating how with the removal of the 'frame' tag from element searches could result in a "NoSuchFrameException" error if multiple elements have the same name attribute.  Also demonstrated is how one can use xpath to search for either frame or iframe or both in such a case as a work around.